Created ReusableZapButton with common zap button logic ZapDVMButton refactored to use ReusableZapButton with DVM-specific configuration
